# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  router.ErrorResponse:
    properties:
      error:
        type: string
      request_id:
        type: string
    type: object
  router.FastDLConfigResponse:
    properties:
      directory:
        type: string
      enabled:
        type: boolean
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the FeatherWings daemon.
  title: FeatherWings API
  version: "1.0"
paths:
...
  /api/servers/{server}/fastdl:
    get:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.FastDLConfigResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Get server FastDL configuration
      tags:
      - Servers
    put:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: FastDL configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/router.FastDLConfigResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.FastDLConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Update server FastDL configuration
      tags:
      - Servers
  /api/servers/{server}/fastdl/disable:
    post:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.FastDLConfigResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Disable FastDL for server
      tags:
      - Servers
  /api/servers/{server}/fastdl/enable:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: FastDL configuration (directory optional)
        in: body
        name: config
        schema:
          $ref: '#/definitions/router.FastDLConfigResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.FastDLConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Enable FastDL for server
      tags:
      - Servers
...
  /download/backup:
    get:
      parameters:
      - description: Signed backup token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - ServerJWT: []
      summary: Download server backup
      tags:
      - Downloads
  /download/file:
    get:
      parameters:
      - description: Signed file token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - ServerJWT: []
      summary: Download server file
      tags:
      - Downloads
...
schemes:
- https
- http
securityDefinitions:
  NodeToken:
    description: 'Supply the node''s bearer token from `config.yml` using the `Authorization:
      Bearer <token>` header.'
    in: header
    name: Authorization
    type: apiKey
  ServerJWT:
    description: Signed JWTs issued by the Panel for server-scoped operations (uploads,
      downloads, websockets). Pass in the `token` query parameter.
    in: query
    name: token
    type: apiKey
swagger: "2.0"